+++ CHANGELOG.md
@@ -15,6 +15,10 @@
 
         ./bin/rake stack:commands:execute_recipes_on_layers layers="Admin, Workers, Engage" recipes="mh-opsworks-recipes::rsyslog-to-loggly"
 
+* *OPTIONAL EDITS TO THE CLUSTER CONFIG*:
+  Recipes, etc for setting up Analytics layer/node. See README.analytics.md in 
+  mh-opsworks for setup instructions.
+
 ## 1.1.6 - 3/3/2016
 
 * *REQUIRES EDITS TO THE CLUSTER CONFIG* *REQUIRES MANUAL CHEF RECIPE RUNS*: 
@@ -48,7 +52,6 @@
   [mh-opsworks](https://github.com/harvard-dce/mh-opsworks/blob/master/README.zadara.md)
   for more information about how to use this recipe. This recipe is independent
   of deployments.
-
 
 * *REQUIRES MANUAL CHEF RECIPE RUNS:* Set the bash prompt to be in color and
   put the cluster name in there. When a cluster includes "production" in the